+++ envoy-control-core/src/main/kotlin/pl/allegro/tech/servicemesh/envoycontrol/snapshot/resource/listeners/filters/JwtFilterFactory.kt
@@ -27,20 +27,29 @@ class JwtFilterFactory(
     private val properties: JwtFilterProperties
 ) {
 
-    private val jwtProviders: Map<ProviderName, JwtProvider> = getJwtProviders()
+    private val jwtProviders: Map<ProviderName, JwtProvider> = getJwtProviders(failedStatusInMetadataEnabled = false)
+    private val jwtProvidersWithJwtStatusMetadata: Map<ProviderName, JwtProvider> =
+        getJwtProviders(failedStatusInMetadataEnabled = true)
     private val clientToOAuthProviderName: Map<String, String> =
         properties.providers.entries.flatMap { (providerName, provider) ->
             provider.matchings.keys.map { client -> client to providerName }
         }.toMap()
 
     fun createJwtFilter(group: Group): HttpFilter? {
+        val selectedJwtProviders =
+            if (group.listenersConfig?.addJwtFailureStatus != false && properties.failedStatusInMetadataEnabled) {
+                jwtProvidersWithJwtStatusMetadata
+            } else {
+                jwtProviders
+            }
+
         return if (shouldCreateFilter(group)) {
             HttpFilter.newBuilder()
                 .setName("envoy.filters.http.jwt_authn")
                 .setTypedConfig(
                     Any.pack(
                         JwtAuthentication.newBuilder().putAllProviders(
-                            jwtProviders
+                            selectedJwtProviders
                         )
                             .addAllRules(createRules(group.proxySettings.incoming.endpoints))
                             .build()
@@ -61,27 +70,34 @@ class JwtFilterFactory(
     private fun containsClientsWithSelector(it: IncomingEndpoint) =
         clientToOAuthProviderName.keys.intersect(it.clients.map { it.name }).isNotEmpty()
 
-    private fun getJwtProviders(): Map<ProviderName, JwtProvider> =
+    private fun getJwtProviders(failedStatusInMetadataEnabled: Boolean): Map<ProviderName, JwtProvider> =
         properties.providers.entries.associate {
-            it.key to createProvider(it.value)
+            it.key to createProvider(it.value, failedStatusInMetadataEnabled)
         }
 
-    private fun createProvider(provider: OAuthProvider) = JwtProvider.newBuilder()
-        .setRemoteJwks(
-            RemoteJwks.newBuilder().setHttpUri(
-                HttpUri.newBuilder()
-                    .setUri(provider.jwksUri.toString())
-                    .setCluster(provider.clusterName)
-                    .setTimeout(
-                        Durations.fromMillis(provider.connectionTimeout.toMillis())
-                    ).build()
+    private fun createProvider(provider: OAuthProvider, failedStatusInMetadataEnabled: Boolean): JwtProvider {
+        val jwtProvider = JwtProvider.newBuilder()
+            .setRemoteJwks(
+                RemoteJwks.newBuilder().setHttpUri(
+                    HttpUri.newBuilder()
+                        .setUri(provider.jwksUri.toString())
+                        .setCluster(provider.clusterName)
+                        .setTimeout(
+                            Durations.fromMillis(provider.connectionTimeout.toMillis())
+                        ).build()
+                )
+                    .setCacheDuration(Durations.fromMillis(provider.cacheDuration.toMillis()))
             )
-                .setCacheDuration(Durations.fromMillis(provider.cacheDuration.toMillis()))
-        )
-        .setForward(properties.forwardJwt)
-        .setForwardPayloadHeader(properties.forwardPayloadHeader)
-        .setPayloadInMetadata(properties.payloadInMetadata)
-        .build()
+            .setForward(properties.forwardJwt)
+            .setForwardPayloadHeader(properties.forwardPayloadHeader)
+            .setPayloadInMetadata(properties.payloadInMetadata)
+
+        if (failedStatusInMetadataEnabled) {
+            jwtProvider.setFailedStatusInMetadata(properties.failedStatusInMetadata)
+        }
+
+        return jwtProvider.build()
+    }
 
     private fun createRules(endpoints: List<IncomingEndpoint>): Set<RequirementRule> {
         return endpoints.flatMap(this::createRulesForEndpoint).toSet()