+++ envoy-control-tests/src/main/kotlin/pl/allegro/tech/servicemesh/envoycontrol/config/consul/ConsulExtension.kt
@@ -5,6 +5,7 @@ import org.junit.jupiter.api.extension.AfterEachCallback
 import org.junit.jupiter.api.extension.BeforeAllCallback
 import org.junit.jupiter.api.extension.ExtensionContext
 import org.testcontainers.containers.Network
+import pl.allegro.tech.servicemesh.envoycontrol.logger
 
 class ConsulExtension : BeforeAllCallback, AfterAllCallback, AfterEachCallback {
 
@@ -16,15 +17,17 @@ class ConsulExtension : BeforeAllCallback, AfterAllCallback, AfterEachCallback {
     }
 
     val server = SHARED_CONSUL
+    private val logger by logger()
     private var started = false
 
     override fun beforeAll(context: ExtensionContext) {
         if (started) {
             return
         }
-
+        logger.info("Consul extension is starting.")
         server.container.start()
         started = true
+        logger.info("Consul extension started.")
     }
 
     override fun afterEach(context: ExtensionContext) {