+++ CHANGELOG.md
@@ -3,6 +3,22 @@
 Lists all changes with user impact.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
+## [0.20.19]
+### Changed
+- Added http compression filter properties to node metadata
+- 
+## [0.20.18]
+### Changed
+- Added http compression filter configuration
+
+## [0.20.17]
+### Fixed
+- Fix JWT provider configuration to not impact lds cache
+- Add missing methods in lua scripts to remove logs about it
+
+## [0.20.16]
+### Changed
+- Add JWT failure reason to metadata and use it in jwt-status field on denied requests
 
 ## [0.21.0]
 ### Changed